fix(mocks): move SymbolicBackend to backends
