+++ src/orquestra/quantum/backends/symbolic_backend.py
@@ -2,12 +2,13 @@
 # © Copyright 2020-2022 Zapata Computing Inc.
 ################################################################################
 from orquestra.quantum.api.backend import QuantumBackend
-from orquestra.quantum.backends.symbolic_simulator import SymbolicSimulator
 from orquestra.quantum.circuits import Circuit
 from orquestra.quantum.measurements import Measurements
 
+from .symbolic_simulator import SymbolicSimulator
 
-class MockQuantumBackend(QuantumBackend):
+
+class SymbolicBackend(QuantumBackend):
 
     supports_batching = False
 
@@ -18,6 +19,6 @@ class MockQuantumBackend(QuantumBackend):
     def run_circuit_and_measure(
         self, circuit: Circuit, n_samples: int, **kwargs
     ) -> Measurements:
-        super(MockQuantumBackend, self).run_circuit_and_measure(circuit, n_samples)
+        super(SymbolicBackend, self).run_circuit_and_measure(circuit, n_samples)
 
         return self._simulator.run_circuit_and_measure(circuit, n_samples)
