+++ src/orquestra/quantum/testing/__init__.py
@@ -2,4 +2,3 @@
 # © Copyright 2020-2022 Zapata Computing Inc.
 ################################################################################
 from ._initialize import *  # noqa: F403
-from .mocks import MockQuantumBackend
