+++ tests/orquestra/quantum/backends/symbolic_simulator_test.py
@@ -10,7 +10,7 @@ from orquestra.quantum.api.backend_test import (
     QuantumSimulatorGatesTest,
     QuantumSimulatorTests,
 )
-from orquestra.quantum.backends.symbolic_simulator import SymbolicSimulator
+from orquestra.quantum.backends import SymbolicSimulator
 from orquestra.quantum.circuits import (
     CNOT,
     RX,
