+++ tests/orquestra/quantum/wavefunction_test.py
@@ -10,7 +10,7 @@ import numpy as np
 import pytest
 from sympy import I, Matrix, Symbol, cos, exp, sin
 
-from orquestra.quantum.backends.symbolic_simulator import SymbolicSimulator
+from orquestra.quantum.backends import SymbolicSimulator
 from orquestra.quantum.circuits._builtin_gates import RX, RY, U3, H, X
 from orquestra.quantum.circuits._circuit import Circuit
 from orquestra.quantum.testing import create_random_wavefunction
